backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
pyyaml==6.0.1
razorpay==1.4.1
paho-mqtt==1.6.1
requests==2.31.0
